@@ -9,14 +9,14 @@
     <packaging>jar</packaging>
 
     <properties>
-        <java.version>11</java.version>
-        <vaadin.version>14.7.1</vaadin.version>
+        <java.version>17</java.version>
+        <vaadin.version>14.8.1</vaadin.version>
     </properties>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.4.5</version>
+        <version>2.5.8</version>
     </parent>
 
     <repositories>
@@ -171,14 +171,14 @@
 		<dependency>
 			<groupId>org.vaadin.crudui</groupId>
 			<artifactId>crudui</artifactId>
-			<version>4.3.2</version>
+			<version>4.4.1</version>
 		</dependency>
 		
 		<!-- Charts -->
 		<dependency>
 			<groupId>com.github.appreciated</groupId>
 			<artifactId>apexcharts</artifactId>
-			<version>2.0.0-beta12</version>
+			<version>2.0.0-beta13</version>
 		</dependency>
 
         <dependency>
@@ -211,7 +211,7 @@
         <dependency>
             <groupId>io.github.bonigarcia</groupId>
             <artifactId>webdrivermanager</artifactId>
-            <version>3.8.1</version>
+            <version>5.0.3</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
