@@ -16,9 +16,9 @@ import org.springframework.context.annotation.PropertySource;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-import se.danielmartensson.entities.LxFlow;
 import se.danielmartensson.entities.LxCurrent;
 import se.danielmartensson.entities.LxData;
+import se.danielmartensson.entities.LxFlow;
 import se.danielmartensson.entities.RsqData;
 import se.danielmartensson.lists.CompleteDataMeasurement;
 import se.danielmartensson.repositories.LxDataRepository;
@@ -29,26 +29,26 @@ import se.danielmartensson.tools.GetSetClassInformation;
 @PropertySource("classpath:schedule.properties")
 public class Schedule {
 
-	public static final String VALVE_TYPE_RSQ = "rsq";	
-	public static final int LENGTH_OF_NOT_CURREPTED_FILE = 1072;
+	public static final String VALVE_TYPE_RSQ = "rsq";
+	public static final int LENGTH_OF_NOT_CURREPTED_FILE = 1077;
 
 	private Logger logger = LoggerFactory.getLogger(getClass());
 
 	@Value("${schedule.testBenchPathCSVFolder}")
 	private String testBenchPathCSVFolder;
-	
+
 	@Value("${schedule.scheduleEnable}")
 	private boolean scheduleEnable;
-	
+
 	@Autowired
 	private LxDataRepository lxDataRepository;
-	
+
 	@Autowired
 	private RsqDataRepository rsqDataRepository;
 
 	@Autowired
 	private FTPConnection ftpConnection;
-	
+
 	@Scheduled(fixedDelayString = "${schedule.testBenchIntervall}")
 	public void testBenchDownloader() {
 		// Check if we should run this schedule.
@@ -66,7 +66,7 @@ public class Schedule {
 
 		// Create list
 		File[] listedCSVFiles = folderPath.listFiles();
-		
+
 		// Check
 		if (listedCSVFiles == null) {
 			logger.info("Error: Cannot read file path to the CSV files!");
@@ -96,31 +96,31 @@ public class Schedule {
 					CSVFile.delete();
 					continue;
 				}
-				
+
 				if (dataRow != null) {
 					// Fill data to the complete measurement object
 					CompleteDataMeasurement completeDataMeasurement = new CompleteDataMeasurement();
 					GetSetClassInformation.setFillFields(completeDataMeasurement, dataRow, 1); // Because 0 = "#"
-					
+
 					// Fill the entities and save
 					if(!completeDataMeasurement.getValveType().toLowerCase().contains(VALVE_TYPE_RSQ)){
 						LxData lxData = new LxData();
 						LxFlow lxFlow = new LxFlow();
 						LxCurrent lxCurrent = new LxCurrent();
-						
+
 						GetSetClassInformation.copyFieldsToDestination(completeDataMeasurement, lxData);
 						GetSetClassInformation.copyFieldsToDestination(completeDataMeasurement, lxCurrent);
 						GetSetClassInformation.copyFieldsToDestination(completeDataMeasurement, lxFlow);
-						
-						lxData.setLxCurrent(lxCurrent);	
-						lxData.setLxFlow(lxFlow);		
+
+						lxData.setLxCurrent(lxCurrent);
+						lxData.setLxFlow(lxFlow);
 						lxDataRepository.save(lxData);
-	
+
 					}else {
 						RsqData rsqData = new RsqData();
-						GetSetClassInformation.copyFieldsToDestination(completeDataMeasurement, rsqData);		
+						GetSetClassInformation.copyFieldsToDestination(completeDataMeasurement, rsqData);
 						rsqDataRepository.save(rsqData);
-					}					
+					}
 				}
 				br.close();
 				CSVFile.delete();
@@ -128,12 +128,12 @@ public class Schedule {
 				e.printStackTrace();
 			}
 		}
-		
+
 	}
 
 	/**
 	 * This returns a filter that ends with endPart
-	 * 
+	 *
 	 * @param endPart
 	 * @return
 	 */
