@@ -1,5 +1,7 @@
 package se.danielmartensson.configurations.security;
 
+import org.springframework.stereotype.Component;
+
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.router.BeforeEnterEvent;
 import com.vaadin.flow.server.ServiceInitEvent;
@@ -7,19 +9,17 @@ import com.vaadin.flow.server.VaadinServiceInitListener;
 
 import se.danielmartensson.views.LoginView;
 
-import org.springframework.stereotype.Component;
-
-@Component 
+@Component
 public class ConfigureUIServiceInitListener implements VaadinServiceInitListener {
 
 	/**
-	 * 
+	 *
 	 */
 	private static final long serialVersionUID = 1L;
 
 	@Override
 	public void serviceInit(ServiceInitEvent event) {
-		event.getSource().addUIInitListener(uiEvent -> { 
+		event.getSource().addUIInitListener(uiEvent -> {
 			final UI ui = uiEvent.getUI();
 			ui.addBeforeEnterListener(this::authenticateNavigation);
 		});
@@ -27,7 +27,7 @@ public class ConfigureUIServiceInitListener implements VaadinServiceInitListener
 
 	private void authenticateNavigation(BeforeEnterEvent event) {
 		if (!LoginView.class.equals(event.getNavigationTarget())
-		    && !SecurityUtils.isUserLoggedIn()) { 
+		    && !SecurityUtils.isUserLoggedIn()) {
 			event.rerouteTo(LoginView.class);
 		}
 	}
