@@ -1,14 +1,14 @@
 package se.danielmartensson.configurations.security;
 
-import org.springframework.security.web.savedrequest.HttpSessionRequestCache;
-
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
+import org.springframework.security.web.savedrequest.HttpSessionRequestCache;
 
 public class CustomRequestCache extends HttpSessionRequestCache {
 
 	@Override
-	public void saveRequest(HttpServletRequest request, HttpServletResponse response) { 
+	public void saveRequest(HttpServletRequest request, HttpServletResponse response) {
 		if (!SecurityUtils.isFrameworkInternalRequest(request)) {
 			super.saveRequest(request, response);
 		}
