@@ -1,13 +1,15 @@
 package se.danielmartensson.configurations.security;
 
-import com.vaadin.flow.server.HandlerHelper.RequestType;
-import com.vaadin.flow.shared.ApplicationConstants;
+import java.util.stream.Stream;
+
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 
-import javax.servlet.http.HttpServletRequest;
-import java.util.stream.Stream;
+import com.vaadin.flow.server.HandlerHelper.RequestType;
+import com.vaadin.flow.shared.ApplicationConstants;
 
 public final class SecurityUtils {
 
@@ -15,14 +17,14 @@ public final class SecurityUtils {
         // Util methods only
     }
 
-    static boolean isFrameworkInternalRequest(HttpServletRequest request) { 
+    static boolean isFrameworkInternalRequest(HttpServletRequest request) {
         final String parameterValue = request.getParameter(ApplicationConstants.REQUEST_TYPE_PARAMETER);
         return parameterValue != null
             && Stream.of(RequestType.values())
             .anyMatch(r -> r.getIdentifier().equals(parameterValue));
     }
 
-    static boolean isUserLoggedIn() { 
+    static boolean isUserLoggedIn() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         return authentication != null
             && !(authentication instanceof AnonymousAuthenticationToken)
